+++ agent-did-extensions/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "plugin-sample",
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "postinstall": "flex-plugin pre-script-check"
+  },
+  "dependencies": {
+    "@emotion/css": "^11.1.3",
+    "@emotion/react": "^11.1.5",
+    "@emotion/styled": "^11.1.5",
+    "@material-ui/core": "^4.11.3",
+    "@twilio-paste/core": "^10.14.0",
+    "@twilio/flex-dev-utils": "^5.1.3",
+    "dotenv-webpack": "^7.1.1",
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
+    "react-redux": "^7.2.2",
+    "react-router-dom": "^5.2.0",
+    "redux": "^4.0.5",
+    "styled-components": "^5.3.5",
+    "@twilio/flex-plugin-scripts": "6.0.3"
+  },
+  "devDependencies": {
+    "@twilio/flex-ui": "2.0.0-beta.3",
+    "react-test-renderer": "17.0.2"
+  }
+}
