+++ agent-did-extensions/public/appConfig.example.js
