+++ agent-did-extensions/src/components/AgentExtensions/EmptyView/SyncEmptyView.js
@@ -1,0 +1,13 @@
+import { TBody, Tr, Td } from '@twilio-paste/core';
+
+const EmptyView = () => {
+  return (
+    <TBody>
+      <Tr>
+        <Td>There are no configured Agent extensions!</Td>
+      </Tr>
+    </TBody>
+  );
+};
+
+export default EmptyView;
