+++ agent-did-extensions/src/components/AgentExtensions/LoadingScreen/LoadingScreen.js
@@ -1,0 +1,18 @@
+import { TBody, Td, Th, Tr } from '@twilio-paste/core';
+
+const LoadingScreen = () => {
+  return (
+    <TBody>
+      <Tr>
+        <Th>
+          <Td>Loading...</Td>
+        </Th>
+        <Th></Th>
+        <Th></Th>
+        <Th></Th>
+      </Tr>
+    </TBody>
+  );
+};
+
+export default LoadingScreen;
