+++ agent-did-extensions/src/components/NewExtension/NewExtensionButton/NewExtensionsButton.js
@@ -1,0 +1,11 @@
+import { Button } from '@twilio-paste/core';
+
+const NewExtButton = ({ clickHandler }) => {
+  return (
+    <Button variant="outlined" onClick={clickHandler}>
+      Add New Extension
+    </Button>
+  );
+};
+
+export default NewExtButton;
