+++ agent-did-extensions/src/components/NewExtension/SidePanel/NewExtension.styles.js
