+++ agent-did-extensions/src/components/SideNavigation/SideNavigationIcon.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import { SideLink, Actions } from '@twilio/flex-ui';
+
+const SideNavigationIcon = ({ activeView }) => {
+  const navigateHandler = () => {
+    Actions.invokeAction('NavigateToView', {
+      viewName: 'agent-extensions-view',
+    });
+  };
+
+  return (
+    <SideLink
+      showLabel={true}
+      icon="Directory"
+      iconActive="DirectoryBold"
+      onClick={navigateHandler}
+      isActive={activeView === 'agent-extensions-view'}
+    />
+  );
+};
+
+export default SideNavigationIcon;
