+++ agent-did-extensions/src/index.js
