+++ agent-did-extensions/src/utils/notificationsUtil.js
