+++ agent-did-extensions/src/utils/workerUtil.js
@@ -1,0 +1,8 @@
+import { Manager } from '@twilio/flex-ui';
+
+const manager = Manager.getInstance();
+const { roles } = manager.user;
+
+export const isAgentRole = () => {
+  return roles.indexOf('agent') >= 0;
+};
