+++ agent-did-extensions/webpack.config.js
