+++ agent-did-extensions/webpack.dev.js
