--- plugin-v1/plugin/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "plugin-sample",
-  "version": "0.0.0",
-  "private": true,
-  "scripts": {
-    "postinstall": "flex-plugin pre-script-check"
-  },
-  "dependencies": {
-    "@twilio/flex-dev-utils": "^5.1.1",
-    "@twilio/flex-plugin-scripts": "5.1.2",
-    "dotenv-webpack": "^7.1.1",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2"
-  },
-  "devDependencies": {
-    "@twilio/flex-ui": "^1",
-    "react-test-renderer": "16.5.2"
-  }
-}
