--- plugin-v1/plugin/public/appConfig.example.js
@@ -1,9 +1,0 @@
-var appConfig = {
-  pluginService: {
-    enabled: true,
-    url: '/plugins',
-  },
-  ytica: false,
-  logLevel: 'info',
-  showSupervisorDesktopView: true,
-};
