--- plugin-v1/plugin/src/.env.example
@@ -1,1 +1,0 @@
-REACT_APP_SYNC_MAP_NAME=
