--- plugin-v1/plugin/src/components/AgentExtensions/AgentExtensions.style.js
@@ -1,11 +1,0 @@
-import styled from 'react-emotion';
-import { TableCell } from '@material-ui/core';
-
-export const TableHeaderCell = styled(TableCell)`
-  height: 60px;
-  font-size: 14px;
-`;
-export const TableData = styled('div')`
-  font-size: 12px;
-  float: left;
-`;
