--- plugin-v1/plugin/src/index.js
@@ -1,4 +1,0 @@
-import * as FlexPlugin from "@twilio/flex-plugin";
-import AgentExtensions from "./AgentExtensions";
-
-FlexPlugin.loadPlugin(AgentExtensions);
