--- plugin-v1/plugin/webpack.config.js
@@ -1,7 +1,0 @@
-module.exports = (config, { isProd, isDev, isTest }) => {
-  /**
-   * Customize the webpack dev-server by modifying the config object.
-   * Consult https://webpack.js.org/configuration/dev-server for more information.
-   */
-  return config;
-};
