--- plugin-v2/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "plugin-sample",
-  "version": "0.0.0",
-  "private": true,
-  "scripts": {
-    "postinstall": "flex-plugin pre-script-check"
-  },
-  "dependencies": {
-    "@emotion/css": "^11.10.0",
-    "@emotion/react": "^11.10.0",
-    "@emotion/styled": "^11.10.0",
-    "@material-ui/core": "^4.12.4",
-    "@twilio-paste/core": "^10.14.0",
-    "@twilio/flex-dev-utils": "^5.1.3",
-    "@twilio/flex-plugin-scripts": "5.1.2",
-    "dotenv-webpack": "^7.1.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.8",
-    "react-router-dom": "^5.3.3",
-    "redux": "^4.2.0",
-    "styled-components": "^5.3.5"
-  },
-  "devDependencies": {
-    "@twilio/flex-ui": "2.0.0-beta.1",
-    "react-test-renderer": "^17.0.2"
-  }
-}
