--- plugin-v2/src/components/AgentExtensions/AgentExtensions.style.js
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-import { Th, Td } from '@twilio-paste/core';
-
-export const TableHeaderCell = styled(Th)`
-  width: "size4000"
-  height: 60px;
-  font-size: 14px;
-`;
-export const TableData = styled(Td)`
-  font-size: 12px;
-  float: left;
-`;
