--- plugin-v2/src/components/utils/workerUtil.js
@@ -1,8 +1,0 @@
-import { Manager } from '@twilio/flex-ui';
-
-const manager = Manager.getInstance();
-
-export const isAgentRole = () => {
-  const { roles } = manager.user;
-  return roles.indexOf('agent') >= 0;
-};
